[수정] 디자인 업데이트 반영, 카드 펑 로직 수정
