+++ SOOUM/SOOUM/DesignSystem/Foundations/UIColor+SOOUM.swift
@@ -33,5 +33,5 @@ extension SOOUMStyle where Base == UIColor {
     static let black = UIColor(hex: "#000000")
     
     // Dim
-    static let dim = UIColor(red: 1, green: 1, blue: 1, alpha: 0.7)
+    static let dim = UIColor(red: 0, green: 0, blue: 0, alpha: 0.5)
 }
