+++ SOOUM/SOOUM/TestTableViewController.swift
@@ -5,13 +5,9 @@
 //  Created by JDeoks on 9/12/24.
 //
 
-import RxSwift
 import UIKit
 
-struct Card {
-    let id: Int
-    let pungTime: Date
-}
+import RxSwift
 
 class TestTableViewController: UITableViewController {
     
@@ -64,21 +60,12 @@ class TestTableViewController: UITableViewController {
         let card = cards[indexPath.row]
         /// 펑타임 설정
         cell.setData(card: card)
-        print(cell.id, "\(String(describing: cell.card?.id)) 구독 됨")
-        cell.didpung
-            .subscribe { _ in
-                if let cardIndex = self.cards.firstIndex(where: { $0.id == card.id }) {
-                    self.cards.remove(at: cardIndex)
-                    // TODO: - 행 삭제 오류 너무 남... 일단 임시로 리로드
-//                    tableView.deleteRows(
-//                        at: [IndexPath(row: cardIndex, section: 0)],
-//                        with: .fade
-//                    )
-                    tableView.reloadData()
-                }
-                print("카드 \(card.id)가 펑됨")
-            }
-            .disposed(by: cell.disposeBag)
+//        cell.didpung
+//            .subscribe { _ in
+//                cell.showDidPung()
+//                print("카드 \(card.id)가 펑됨")
+//            }
+//            .disposed(by: cell.disposeBag)
         
         return cell
     }
@@ -88,16 +75,5 @@ class TestTableViewController: UITableViewController {
         heightForRowAt indexPath: IndexPath
     ) -> CGFloat {
         (tableView.bounds.width - 40) * 0.9 + 10
-    }
-    
-    override func tableView(
-        _ tableView: UITableView,
-        didEndDisplaying cell: UITableViewCell,
-        forRowAt indexPath: IndexPath
-    ) {
-        if let cell = cell as? SOMCardTableViewCell {
-            print(cell.id, "\(String(describing: cell.card?.id)) 구독 해제")
-            cell.disposeBag = DisposeBag()
-        }
     }
 }
